pyproject.toml
[project]
name = "cold-mailing-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.19",
    "pandas>=2.3.0",
    "requests>=2.32.4",
    "schedule>=1.2.2",
    "supabase>=2.16.0",
]
